Code refactoring. Coupon class is now with Lombok.
ATTENTION: when you modify Coupon class on Client you must update also Coupon class on Provider, because GSon (JSON library) uses Coupon.java.

+++ sPFCouponingClient/src/main/java/it/polimi/spf/demo/couponing/client/TriggerIntentReceiver.java
@@ -32,7 +32,7 @@ public class TriggerIntentReceiver extends BroadcastReceiver {
 		Map<String, ?> prefs = context.getSharedPreferences(PREF_FILE, Context.MODE_PRIVATE).getAll();
 		String category = null;
 		for(Map.Entry<String, ?> entry : prefs.entrySet()){
-			if(((Long) entry.getValue()).equals(triggerId)){
+			if((entry.getValue()).equals(triggerId)){
 				category = entry.getKey();
 				break;
 			}
